+++ src/main/java/module-info.java
@@ -1,4 +1,4 @@
-module pihole {
+module APIs {
     requires javafx.controls;
     requires javafx.fxml;
     requires javafx.web;