+++ CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Optimized API Calls.
 - Handle Timeout.
+- Fixed bug if no data in pihole server.
+- Changed Name of service module.
 
 ## Ver [1.5.2] - 02-13-2022